Re-wrote match detection
My first attempt was pretty rubbish... after thinking it through a little I realized that the simplest thing is just to scan the whole board. I can repeat this over and over until no further matches are found. This allows for new matches created by drops to be handled. It's probably not the most efficient solution and even within that there is probably room for optimization; however, for an 8x8 board this brute force approach is probably just as good as any other. No point making it complicated for the sake of it!

The code is still a little ragged and there are some redundancy optimization opportunities to be had... which I'll sort out in a commit coming this way, soon!